+++ src/extensions/views/InstallCustomExtension/components/InstallCustomExtensionFromForm/InstallCustomExtensionFromForm.tsx
@@ -61,6 +61,12 @@ export const InstallCustomExtensionFromForm = ({
     return "issues" in manifestValidation ? manifestValidation.issues : undefined;
   }, [manifest]);
 
+  /**
+   * Prevent installation if validation fails. In the future we can change errors to warnings, but first we need to add special handling
+   * to Dashboard, e.g. render "warning" near widget that widget is broken.
+   */
+  const canInstall = !issues && manifest;
+
   return (
     <>
       <InstallTopNav title={intl.formatMessage(headerTitles.addCustomExtensionManifest)} />
@@ -88,7 +94,7 @@ export const InstallCustomExtensionFromForm = ({
         <Savebar.Spacer />
         <Savebar.CancelButton href={previousPagePath} />
         <Savebar.ConfirmButton
-          disabled={!manifest}
+          disabled={!canInstall}
           transitionState={isSubmittingInstallation ? "loading" : "default"}
           onClick={() => submitInstallApp()}
         >