+++ src/warehouses/components/WarehouseInfo/WarehouseInfo.tsx
@@ -1,4 +1,5 @@
 import { DashboardCard } from "@dashboard/components/Card";
+import { FormSpacer } from "@dashboard/components/FormSpacer";
 import { WarehouseErrorFragment } from "@dashboard/graphql";
 import { FormChange } from "@dashboard/hooks/useForm";
 import { commonMessages } from "@dashboard/intl";
@@ -8,7 +9,10 @@ import { TextField } from "@material-ui/core";
 import { useIntl } from "react-intl";
 
 interface WarehouseInfoProps {
-  data: Record<"name", string>;
+  data: {
+    name: string;
+    email: string;
+  };
   disabled: boolean;
   errors: WarehouseErrorFragment[];
   onChange: FormChange;
@@ -16,7 +20,7 @@ interface WarehouseInfoProps {
 
 const WarehouseInfo = ({ data, disabled, errors, onChange }: WarehouseInfoProps) => {
   const intl = useIntl();
-  const formErrors = getFormErrors(["name"], errors);
+  const formErrors = getFormErrors(["name", "email"], errors);
 
   return (
     <DashboardCard data-test-id="general-information-section" paddingTop={9}>
@@ -43,6 +47,25 @@ const WarehouseInfo = ({ data, disabled, errors, onChange }: WarehouseInfoProps)
             inputProps: {
               autoComplete: "none",
             },
+          }}
+        />
+        <FormSpacer />
+        <TextField
+          disabled={disabled}
+          error={!!formErrors.email}
+          data-test-id="company-email-input"
+          fullWidth
+          helperText={getWarehouseErrorMessage(formErrors.email, intl)}
+          label={intl.formatMessage({
+            id: "sy+pv5",
+            defaultMessage: "Email",
+          })}
+          name={"email"}
+          value={data.email}
+          onChange={onChange}
+          InputProps={{
+            autoComplete: "email",
+            spellCheck: false,
           }}
         />
       </DashboardCard.Content>